added  tag service verification

+++ src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -2,7 +2,10 @@ package com.epam.esm.service.impl;
 
 import com.epam.esm.dao.TagDAO;
 import com.epam.esm.entity.Tag;
+import com.epam.esm.exception.TagIdException;
+import com.epam.esm.exception.TagNameException;
 import com.epam.esm.service.TagService;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -21,6 +24,9 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public long create(Tag tag) {
+        if(!StringUtils.isAlpha(tag.getName())){
+            throw  new TagNameException(tag.getName());
+        }
         return tagDAO.create(tag);
     }
 
@@ -36,11 +42,15 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void update(Map<String, Object> updates) {
+        if
         tagDAO.update(updates);
     }
 
     @Override
     public void delete(long id) {
+        if(id<=0||id>Long.MAX_VALUE){
+            throw new TagIdException(id);
+        }
         tagDAO.deleteById(id);
     }
 
@@ -52,5 +62,10 @@ public class TagServiceImpl implements TagService {
     @Override
     public boolean existByName(String name) {
         return tagDAO.existByName(name);
+    }
+
+    @Override
+    public Tag findByName(String name) {
+        return tagDAO.findByName(name);
     }
 }
